+++ CHANGES.rst
@@ -6,7 +6,9 @@ CHANGES
 2.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Errors executing SQL statements now doom the entire transaction,
+  causing `transaction.interfaces.DoomedTransaction` exception on
+  any attempts to commit it.  A failed transaction must be aborted.
 
 
 2.0.1 (2020-10-13)